#5513 - Created DocumentList
